requirements.txt
torch
torchvision
torchaudio
torchinfo
tqdm
pillow
matplotlib
pyyaml
black
opencv-python
